\name{elnglm-package}
\alias{elnglm-package}
\alias{elnglm}
\docType{package}
\title{
\packageTitle{elnglm}
}
\description{
\packageDescription{elnglm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{elnglm}
\packageIndices{elnglm}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{elnglm}

Maintainer: \packageMaintainer{elnglm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
