% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnglm.R
\name{glmDataGen}
\alias{glmDataGen}
\title{Data generator under the framework of generalized linear model}
\usage{
glmDataGen(
  n,
  d,
  family = c("gaussian", "binomial", "multinomial"),
  trueb0,
  trueb,
  s = 0.5,
  seed = NULL
)
}
\arguments{
\item{n}{integer. The number of observations.}

\item{d}{integer. The number of predictors.}

\item{family}{string. The value of starting inertia weight in PSO updating procedure. The default is 1.2.}

\item{trueb0}{double.}

\item{trueb}{vector.}

\item{s}{double.}

\item{seed}{integer. The random seed.}
}
\value{
An List.
\itemize{
\item{x}{ the matrix of the predictors.}
\item{y}{ the vector of the response variable. 
For \code{family = 'multinomial'}, the output is a matrix of size $n\times$#(categories).}
}
}
\description{
Data generator under the framework of generalized linear model
}
\examples{
# Intercept
trueb0 <- 1
# Regression Coefficients (the first 3 are active)
trueact <- c(1, 1, 1, 0, 0, 0, 0, 0, 0, 0)
trueb <- runif(10, -1, 1)*10
trueb[which(trueact == 0)] <- 0 

# Generate data of continuous response
df <- glmDataGen(n = 500, d = 10, family = "gaussian", trueb0, trueb, s = 0.5, seed = 1)

# Generate data of binary response
dfb <- glmDataGen(n = 500, d = 10, family = "binomial", trueb0, trueb, seed = 1)

# Generate data of multi-categorical response (not run)
#
# Intercepts of the three log-linear models 
# trueb0 <- c(1, 1, 1)
# Regression Coefficients of the three log-linear models 
# trueact <- cbind(
#   c(1, 1, 1, 0, 0, 0, 0, 0, 0, 0),
#   c(0, 0, 0, 1, 1, 1, 1, 0, 0, 0),
#   c(0, 0, 0, 0, 0, 1, 1, 1, 1, 0)
# )
# trueb <- matrix(runif(10*3, -1, 1)*10, 10, 3)
# for (m in 1:3) { trueb[which(trueact[,m] == 0),m] <- 0 }
# 
# Generate data. The response y is a matrix of 3 columns.
# dfm <- glmDataGen(n = 500, d = 10, family = "multinomial", trueb0, trueb, seed = 1)

}
